simplified code and PHP version checks
