+++ src/DoctrineObject.php
@@ -40,7 +40,6 @@ use function is_callable;
 use function is_int;
 use function is_object;
 use function is_string;
-use function interface_exists;
 use function method_exists;
 use function property_exists;
 use function sprintf;
@@ -316,7 +315,8 @@ class DoctrineObject extends AbstractHydrator
             return null;
         }
 
-        if (PHP_VERSION_ID >= 80100 && interface_exists(BackedEnum::class) && $value instanceof BackedEnum) {
+        // BackedEnum is available from PHP 8.1 on
+        if ($value instanceof BackedEnum) {
             return $value;
         }
