+++ tests/Assets/SimpleEnumPhp81.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DoctrineTest\Laminas\Hydrator\Assets;
+
+enum SimpleEnumPhp81: int
+{
+case One = 1;
+case Two = 2;
+    }
